requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
requests
python-dotenv>=0.5.1
pypandoc
pdftotext
numpy
pandas
